feat: connect vercel image to token seo
