+++ app/collect/[collection]/[tokenId]/page.tsx
@@ -1,28 +1,26 @@
-"use client";
-
-import { ZoraChains } from "@/lib/zora/zoraToViem";
-import { ZORA_TO_VIEM } from "@/lib/zora/zoraToViem";
-import { Address } from "viem";
-import { useParams } from "next/navigation";
-import * as chains from "viem/chains";
-import { CollectionProvider } from "@/providers/CollectionProvider";
 import TokenPage from "@/components/TokenPage";
+import { Metadata, NextPage } from "next";
 
-export default function Token() {
-  const params = useParams();
-  const collection = params.collection as string;
-  const tokenId = params.tokenId as string;
-  const [chain, address] = collection.split("%3A");
-  const viemChainName = ZORA_TO_VIEM[chain as ZoraChains];
-  const viemChain = chains[viemChainName];
+type Props = {
+  params: Promise<{ collection: string; tokenId: string }>;
+};
 
-  return (
-    <CollectionProvider chainId={viemChain.id} address={address as Address}>
-      <TokenPage
-        chainId={viemChain.id}
-        address={address as Address}
-        tokenId={tokenId as Address}
-      />
-    </CollectionProvider>
-  );
+export async function generateMetadata({ params }: Props): Promise<Metadata> {
+  const { tokenId, collection } = await params;
+
+  return {
+    title: "In Process",
+    description: "Imagined by LATASHÁ",
+    openGraph: {
+      title: "In Process",
+      description: "Imagined by LATASHÁ",
+      images: [
+        `https://in-process-seven.vercel.app/api/og/token?collection=${collection}&tokenId=${tokenId}`,
+      ],
+    },
+  };
 }
+
+const Token: NextPage = () => <TokenPage />;
+
+export default Token;
